fix(action-sheet): fix the multiple pages error
